[Chore] #99 - 잘못 분리된 Extension 삭제
