+++ ToDwoong/ToDwoong/Presentation/Map/ViewController/MapViewController.swift
@@ -321,3 +321,20 @@ extension MapViewController {
         present(detailVC, animated: true, completion: nil)
     }
 }
+
+extension MapViewController {
+    func zoomToTodo(_ todo: TodoModel, completion: @escaping () -> Void) {
+        if let annotation = todoAnnotationMap[todo.id?.uuidString ?? ""] {
+            let region = MKCoordinateRegion(center: annotation.coordinate,
+                                            latitudinalMeters: 300,
+                                            longitudinalMeters: 300)
+            customMapView.mapView.setRegion(region, animated: true)
+            
+            DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
+                completion()
+            }
+        } else {
+            completion()
+        }
+    }
+}
